pkg/client/cli: Move filenames around, move things between files
Absolutely no code is changing in this commit.  I'm just organizing
things.
