+++ pkg/client/cli/cmd_license.go
