+++ pkg/client/cli/cmd_license_test.go
