+++ pkg/client/cli/cmd_preview.go
