+++ pkg/client/cli/cmd_status.go
