+++ pkg/client/cli/cmd_uninstall.go
