+++ pkg/client/cli/util.go
@@ -14,6 +14,22 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/cliutil"
 )
 
+// quit sends the quit message to the daemon and waits for it to exit.
+func quit(ctx context.Context) error {
+	// When the daemon shuts down, it will tell the connector to shut down.
+	if err := cliutil.QuitDaemon(ctx); err != nil {
+		return err
+	}
+
+	// But also do that ourselves; to ensure the connector is killed even if daemon isn't
+	// running.  If the daemon already shut down the connector, then this is a no-op.
+	if err := cliutil.QuitConnector(ctx); err != nil {
+		return err
+	}
+
+	return nil
+}
+
 func kubeFlagMap() map[string]string {
 	kubeFlagMap := make(map[string]string)
 	kubeFlags.VisitAll(func(flag *pflag.Flag) {
@@ -93,23 +109,4 @@ func setConnectInfo(ctx context.Context, stdout io.Writer) (*connector.ConnectIn
 		return nil, err
 	}
 	return resp, nil
-}
-
-func connectCommand() *cobra.Command {
-	return &cobra.Command{
-		Use:  "connect [flags] [-- <command to run while connected>]",
-		Args: cobra.ArbitraryArgs,
-
-		Short: "Connect to a cluster",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) == 0 {
-				return withConnector(cmd, true, func(_ context.Context, _ connector.ConnectorClient, _ *connector.ConnectInfo) error {
-					return nil
-				})
-			}
-			return withConnector(cmd, false, func(ctx context.Context, _ connector.ConnectorClient, _ *connector.ConnectInfo) error {
-				return start(ctx, args[0], args[1:], true, cmd.InOrStdin(), cmd.OutOrStdout(), cmd.ErrOrStderr())
-			})
-		},
-	}
 }
